algorithms/pyproject.toml
[project]
name = "chainwhales-algo"
version = "0.1.0"
description = "Phase 0 prototype for ChainWhales accumulation detection"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "ChainWhales"}
]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.8.2",
  "tenacity>=8.5.0",
  "pyyaml>=6.0.2"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.5",
  "pytest>=8.3.3"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

